backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-texttospeech==2.16.3
python-multipart==0.0.6
pydantic==2.10.0
python-dotenv==1.1.1